Refactor: Theming and Data Access Improvements, Project Structure Updates Theming System Overhaul • Introduced AppThemes Class: Added a new AppThemes class to centralize theme definitions and management. This class provides a registry of available themes, exposes theme property accessors, and supports querying and fallback logic for theme selection. • Removed Internal AppTheme from DgvDesigner: The AppTheme class is now defined in AppThemes.cs and is no longer an internal class within DgvDesigner, promoting better separation of concerns and reusability. • Enhanced DgvDesigner.cs: • Refactored to utilize the new AppThemes class for all theme-related operations. • Improved XML documentation for all public methods and regions. • Added/updated helper methods for theme property access, theme existence checks, and test support. • Improved recursive theming logic for forms and controls, including DataGridView, Button, TabControl, TextBox, and Label. • Ensured theming respects custom font sizes and updates form titles to reflect the active theme. • Changed ThemeFontSize in WipAppVariables.cs: Updated the type from int to float to allow for more precise font size configuration. Data Access Refactoring • Introduced MySqlDataAdapterWrapper: Created a new MySqlDataAdapterWrapper class implementing the IDataAdapter interface, enabling flexible and testable data access abstraction over MySQL operations. • Refactored Data Access in Forms: • AdvancedInventoryEntryForm.cs: Replaced direct usage of MySqlDataAdapter with the new MySqlDataAdapterWrapper for improved abstraction and testability. • MainFormComboBoxDataHelper.cs: Updated all data access logic to use the IDataAdapter interface, decoupling the code from specific data adapter implementations and improving maintainability. Project and Solution Structure • Project File Updates (MTM_WIP_Application.csproj): • Updated target framework to .NET 9.0 (net9.0-windows7.0). • Added a reference to Microsoft.TestPlatform for enhanced unit testing support. • Cleaned up and reorganized compile/remove/include directives for better clarity. • Unit Testing Enhancements: • Added a new unit test project to the solution for improved test coverage. • Updated the solution file to include the new test project. • Enhanced testability of startup logic by refactoring Program.cs to extract and expose startup routines. • Assembly Info Update: • Modified AssemblyInfo.cs to update the InternalsVisibleTo attribute, ensuring correct visibility for the new or updated unit test project. Additional Improvements • DgvPrinter.cs: Added comprehensive XML documentation to improve maintainability and developer understanding. • General Code Quality: • Improved XML documentation across multiple files. • Ensured all new and refactored code adheres to the project’s coding standards and .NET 9.0 best practices.

+++ Core/WipAppVariables.cs
@@ -41,7 +41,7 @@ internal static class WipAppVariables
     public static string? RemoveUser; // User performing removal
 
     // Theme font size (added from solution context)
-    public static int ThemeFontSize = 9;
+    public static float ThemeFontSize = 9f;
     public static string? TransferType; // Type of transfer
 
     // Current user (resolved at startup)
@@ -66,6 +66,6 @@ internal static class WipAppVariables
     // Operation used for removal, with internal setter
     public static string? RemoveOperation { get; internal set; }
 
-// Connection string for database access
+    // Connection string for database access
     public static string ConnectionString = SqlVariables.GetConnectionString(null, null, null, null);
 }
